+++ tell-agent/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.3.0](https://github.com/omercnet/paseo-plugins/compare/tell-agent-v0.2.0...tell-agent-v0.3.0) (2026-09-18)
+
+
+### Features
+
+* **release:** publish plugins to npm ([#80](https://github.com/omercnet/paseo-plugins/issues/80)) ([3c93048](https://github.com/omercnet/paseo-plugins/commit/3c93048cfefda97d8c2bc1631e3428fb64bdad09))
+
+
+### Bug Fixes
+
+* **tell-agent:** support Paseo 0.9 beta ([#85](https://github.com/omercnet/paseo-plugins/issues/85)) ([f8425f6](https://github.com/omercnet/paseo-plugins/commit/f8425f6c4b6271db5514abe3bf3c541c6240ccef))
+
 ## [0.2.0](https://github.com/omercnet/paseo-plugins/compare/tell-agent-v0.1.0...tell-agent-v0.2.0) (2026-09-10)